bugfix: even core plugins couldn't load after config init
