+++ src/yodapa/plugins/dev.py
@@ -1,0 +1,10 @@
+import os
+
+import typer
+
+app = typer.Typer()
+
+
+@app.command()
+def speedtest():
+    os.system("speedtest-cli")
